+++ examples/vnet_integration/modules.tf
@@ -27,6 +27,40 @@ module "logs" {
   resource_group_name = module.rg.resource_group_name
 }
 
+module "vnet" {
+  source  = "claranet/vnet/azurerm"
+  version = "x.x.x"
+
+  environment    = var.environment
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
+  client_name    = var.client_name
+  stack          = var.stack
+
+  resource_group_name = module.rg.resource_group_name
+  vnet_cidr           = [local.vnet_cidr]
+}
+
+module "subnet" {
+  source  = "claranet/subnet/azurerm"
+  version = "x.x.x"
+
+  for_each = { for subnet in local.subnets : subnet.name => subnet }
+
+  environment    = var.environment
+  location_short = module.azure_region.location_short
+  client_name    = var.client_name
+  stack          = var.stack
+
+  custom_subnet_name = each.key
+
+  resource_group_name  = module.rg.resource_group_name
+  virtual_network_name = module.vnet.virtual_network_name
+  subnet_cidr_list     = each.value.cidr
+
+  service_endpoints = each.value.service_endpoints
+}
+
 module "acr" {
   source  = "claranet/acr/azurerm"
   version = "x.x.x"
@@ -42,6 +76,10 @@ module "acr" {
   logs_destinations_ids = [
     module.logs.logs_storage_account_id,
     module.logs.log_analytics_workspace_id
+  ]
+
+  allowed_subnets = [
+    for s in module.subnet : s.subnet_id
   ]
 
   extra_tags = {