+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+Added
+  * AZ-504: Adding diagnostic logging, trust policies & retention options.
+  * AZ-504: Adding Network rules (ip_rules and authorized subnets)
+
 Changed
   * AZ-572: Revamp examples and improve CI